requirements.txt
astroid==2.3.3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
gitdb==4.0.5
GitPython==3.1.14
idna==2.8
isort==4.3.21
Jinja2==2.10.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.4.4
python-gitlab==1.8.0
requests==2.22.0
six==1.12.0
smmap==3.0.5
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
